requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
astropy==3.2.1
atomicwrites==1.3.0
attrs==19.1.0
bleach==3.1.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
confuse==1.0.0
cryptography==2.7
cycler==0.10.0
decorator==4.4.0
defusedxml==0.5.0
entrypoints==0.3
htmlmin==0.1.12
idna==2.8
importlib-metadata==0.18
ipython-genutils==0.2.0
isort==4.3.21
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.2
jupyter-client==5.3.1
jupyter-core==4.4.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
lightgbm==2.2.3
llvmlite==0.29.0
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
missingno==0.4.2
mistune==0.8.4
more-itertools==7.2.0
nbconvert==5.6.0
nbformat==4.4.0
numba==0.45.1
numpy==1.17.0
packaging==19.0
pandas==0.25.0
pandas-profiling==2.3.0
pandocfilters==1.4.2
patsy==0.5.1
phik==0.9.8
pluggy==0.12.0
pprint==0.1
psutil==5.6.3
psycopg2==2.8.3
py==1.8.0
pycparser==2.19
Pygments==2.4.2
pylint==2.3.1
pyOpenSSL==19.0.0
pyparsing==2.4.2
pyrsistent==0.15.4
PySocks==1.7.0
pytest==5.1.0
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.3.0
pytest-openfiles==0.4.0
pytest-pylint==0.14.1
pytest-remotedata==0.3.1
pytest-runner==5.1
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
pyzmq==18.0.2
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.9.0
six==1.12.0
sklearn==0.0
SQLAlchemy==1.3.7
statsmodels==0.10.1
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
wrapt==1.11.2
zipp==0.5.2
